Fixed free() invalid pointers and heap corruption issues when mixing leveldb shared objects
If leveldb was compiled as a shared object, it could be compiled with different compiler options than php-leveldb extension (for example, leveldb in release mode and php-leveldb in debug mode)
which can result in the two using different memory management. Therefore leveldb itself needs to free anything it allocated instead of php-leveldb doing it.

As an example, I encountered heap corruption on Windows with Mojang's leveldb-mcpe when leveldb was compiled in release mode and php-leveldb in debug mode. `free_dbg()` would crash trying to free memory allocated by leveldb, because the corresponding debug functions did not allocate the memory.